# extras/backend_openapi.json
{"openapi":"3.0.2","info":{"title":"BlackRock Backend (orchestrator)","version":"0.1.0"},"paths":{"/health":{"get":{"summary":" Health","operationId":"_health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/transactions":{"post":{"summary":"Create Transaction","operationId":"create_transaction_transactions_post","parameters":[{"required":false,"schema":{"title":"Idempotency-Key","type":"string"},"name":"idempotency-key","in":"header"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransactionIn"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/transactions/local":{"post":{"summary":"Transactions Forward","operationId":"transactions_forward_transactions_local_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransactionReq"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/payout/iso20022":{"post":{"summary":"Payout Iso20022","operationId":"payout_iso20022_payout_iso20022_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ISO20022Payout"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/payout/crypto":{"post":{"summary":"Payout Crypto","operationId":"payout_crypto_payout_crypto_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CryptoPayout"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/dashboard":{"get":{"summary":"Dashboard","operationId":"dashboard_dashboard_get","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}}}},"/monitor":{"get":{"summary":"Monitor","operationId":"monitor_monitor_get","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"CryptoPayout":{"title":"CryptoPayout","required":["to_address","amount"],"type":"object","properties":{"to_address":{"title":"To Address","type":"string"},"amount":{"title":"Amount","type":"number"},"currency":{"title":"Currency","type":"string","default":"ETH"},"correlation_id":{"title":"Correlation Id","type":"string"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"ISO20022Payout":{"title":"ISO20022Payout","required":["creditor_name","amount","pain_xml"],"type":"object","properties":{"creditor_name":{"title":"Creditor Name","type":"string"},"amount":{"title":"Amount","type":"number"},"currency":{"title":"Currency","type":"string","default":"USD"},"pain_xml":{"title":"Pain Xml","type":"string"},"correlation_id":{"title":"Correlation Id","type":"string"}}},"TransactionIn":{"title":"TransactionIn","required":["merchant_id","cardNumber","expiry","cvc","amount","currency","protocol","authCode"],"type":"object","properties":{"merchant_id":{"title":"Merchant Id","type":"string","description":"Merchant identifier"},"cardNumber":{"title":"Cardnumber","type":"string"},"expiry":{"title":"Expiry","type":"string"},"cvc":{"title":"Cvc","type":"string"},"amount":{"title":"Amount","type":"number"},"currency":{"title":"Currency","type":"string"},"protocol":{"title":"Protocol","type":"string"},"authCode":{"title":"Authcode","type":"string"},"payoutMethod":{"title":"Payoutmethod","type":"string"},"payoutDetails":{"title":"Payoutdetails","type":"object"}}},"TransactionReq":{"title":"TransactionReq","required":["cardNumber","expiry","cvc","amount"],"type":"object","properties":{"merchant_id":{"title":"Merchant Id","type":"string"},"cardNumber":{"title":"Cardnumber","type":"string"},"expiry":{"title":"Expiry","type":"string"},"cvc":{"title":"Cvc","type":"string"},"amount":{"title":"Amount","type":"number"},"currency":{"title":"Currency","type":"string","default":"USD"},"protocol":{"title":"Protocol","type":"string"},"authCode":{"title":"Authcode","type":"string"},"txn_id":{"title":"Txn Id","type":"string"},"correlation_id":{"title":"Correlation Id","type":"string"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}